requirements_updateIndex.txt
torch 
torchvision 
torchaudio
annoy
pymongo
python-dotenv
tqdm
joblib
h5py
transformers
nltk